# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Journey API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/image": {
            "get": {
                "tags": [
                    "imageInput"
                ],
                "summary": "List images",
                "description": "Get a list of all image inputs.",
                "operationId": "imageInputList",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ImageInput"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "imageInput"
                ],
                "summary": "Create image",
                "description": "Create a new image input.",
                "operationId": "imageInputCreate",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name of the input.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "The image file.",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageInput"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The entered parameters are not valid."
                    }
                }
            }
        },
        "/image/{id}": {
            "get": {
                "tags": [
                    "imageInput"
                ],
                "summary": "Show image",
                "description": "Get a single image input by id.",
                "operationId": "imageInputShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageInput"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The image input does not exist."
                    }
                }
            },
            "delete": {
                "tags": [
                    "imageInput"
                ],
                "summary": "Delete image",
                "description": "Delete an image input.",
                "operationId": "imageInputDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "The image input does not exist."
                    }
                }
            },
            "patch": {
                "tags": [
                    "imageInput"
                ],
                "summary": "Update image",
                "description": "Update an existing image input.",
                "operationId": "imageInputUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "The image file.",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageInput"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The entered parameters are not valid."
                    },
                    "404": {
                        "description": "The image input does not exist."
                    }
                }
            }
        },
        "/input": {
            "get": {
                "tags": [
                    "input"
                ],
                "summary": "List inputs",
                "description": "Get a list of all inputs.",
                "operationId": "inputList",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Input"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "input"
                ],
                "summary": "Create input",
                "description": "Create a new input.",
                "operationId": "inputCreate",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name of the input.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "input_type_id",
                        "in": "query",
                        "description": "The id of the specialized object.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "input_type_type",
                        "in": "query",
                        "description": "The class name of the specialized object.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Input"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The entered parameters are not valid."
                    }
                }
            }
        },
        "/input/{id}": {
            "get": {
                "tags": [
                    "input"
                ],
                "summary": "Show input",
                "description": "Get a single input by id.",
                "operationId": "inputShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Input"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The input does not exist."
                    }
                }
            },
            "delete": {
                "tags": [
                    "input"
                ],
                "summary": "Delete input",
                "description": "Delete an input.",
                "operationId": "inputDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "The input does not exist."
                    }
                }
            },
            "patch": {
                "tags": [
                    "input"
                ],
                "summary": "Update input",
                "description": "Update an existing input.",
                "operationId": "inputUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name of the input.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "input_type_id",
                        "in": "query",
                        "description": "The id of the specialized object.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "input_type_type",
                        "in": "query",
                        "description": "The class name of the specialized object.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Input"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The entered parameters are not valid."
                    },
                    "404": {
                        "description": "The input does not exist."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ImageInput": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "file_url": {
                        "type": "string",
                        "example": "http://targeturl.com/uploads/images/name.png"
                    },
                    "type": {
                        "type": "string",
                        "example": "png"
                    },
                    "size": {
                        "type": "integer",
                        "example": "3453979"
                    },
                    "created_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    },
                    "updated_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ImageInput"
                }
            },
            "Input": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Take Photo"
                    },
                    "data_type_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "data_type_type": {
                        "type": "string",
                        "example": "App\\\\Models\\\\ImageInput"
                    },
                    "created_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    },
                    "updated_at": {
                        "type": "date",
                        "example": "2022-05-25 22:15:23"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Input"
                }
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}